Time: 0 ms (100.00%) | Memory: 21.4 MB (30.99%) - LeetSync

+++ 450-delete-node-in-a-bst/delete-node-in-a-bst.py
@@ -5,42 +5,28 @@
 #         self.left = left
 #         self.right = right
 class Solution:
-    # One step right and then always left
-    def successor(self, root: TreeNode) -> int:
-            root = root.right
-            while root.left:
-                root = root.left
-            return root.val
+    def successor(self, rt) :
+            rt = rt.right
+            while rt.left: rt = rt.left
+            return rt.val
+    def predecessor(self, rt):
+        rt = rt.left
+        while rt.right: rt = rt.right
+        return rt.val
+    def deleteNode(self, root: Optional[TreeNode], key: int) -> Optional[TreeNode]:
         
-    # One step left and then always right
-    def predecessor(self, root: TreeNode) -> int:
-        root = root.left
-        while root.right:
-            root = root.right
-        return root.val
-
-    def deleteNode(self, root: TreeNode, key: int) -> TreeNode:
-        if not root:
-            return None
-
-        # delete from the right subtree
+        if not root: return None
         if key > root.val:
             root.right = self.deleteNode(root.right, key)
         # delete from the left subtree
         elif key < root.val:
             root.left = self.deleteNode(root.left, key)
-        # delete the current node
         else:
-            # the node is a leaf
-            if not (root.left or root.right):
-                root = None
-            # The node is not a leaf and has a right child
+            if not (root.left or root.right): return None
             elif root.right:
                 root.val = self.successor(root)
                 root.right = self.deleteNode(root.right, root.val)
-            # the node is not a leaf, has no right child, and has a left child    
             else:
                 root.val = self.predecessor(root)
                 root.left = self.deleteNode(root.left, root.val)
-                        
         return root
